move Case to base module
