+++ base/src/org/intellij/grammar/util/Case.java
@@ -1,8 +1,8 @@
 /*
- * Copyright 2011-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+ * Copyright 2011-2026 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
  */
 
-package org.intellij.grammar.generator;
+package org.intellij.grammar.util;
 
 import org.jetbrains.annotations.NotNull;
