+++ bnf-language/src/org/intellij/grammar/generator/CommonRendererUtils.java
@@ -10,6 +10,7 @@ import com.intellij.psi.codeStyle.NameUtil;
 import org.intellij.grammar.KnownAttribute;
 import org.intellij.grammar.psi.BnfFile;
 import org.intellij.grammar.psi.BnfRule;
+import org.intellij.grammar.util.Case;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
